Sir-Aneeq-tasks/task3/health_agent_SDK/pyproject.toml
[project]
name = "health-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.1.0",
    "streamlit>=1.47.0",
]
